% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookUpSingleDisease.R
\name{checkUnspecified}
\alias{checkUnspecified}
\title{Check category coincidence}
\usage{
checkUnspecified(pattern, cat, catalog)
}
\arguments{
\item{pattern}{A string of the name of the disease to look up.}

\item{cat}{A string indicating the category}

\item{catalog}{A data frame containing the catalog where the search will be made.}
}
\value{
A string with the category or subcategory to print
}
\description{
Check category coincidence
}
