% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFunctions.R
\name{cleanCancer}
\alias{cleanCancer}
\title{Clean cancer pattern}
\usage{
cleanCancer(pattern)
}
\arguments{
\item{pattern}{The string whose stopwords and cancer stuff will be removed}
}
\value{
A new version of the string in lowercase without noise.
}
\description{
Clean cancer pattern
}
