% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFunctions.R
\name{cleanString}
\alias{cleanString}
\title{Clean noise from string}
\usage{
cleanString(pattern)
}
\arguments{
\item{pattern}{The string whose noise will be removed (special characters, double spaces, etc.).}
}
\value{
A new version of the string in lowercase without noise.
}
\description{
Clean noise from string
}
