% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICDLookUp.R
\name{ICDLookUp}
\alias{ICDLookUp}
\title{Main function for looking up a given pattern.}
\usage{
ICDLookUp(pattern, nMatches = 1, jwBound = 0.9)
}
\arguments{
\item{pattern}{A string of the name of the disease to look up.}

\item{nMatches}{A number indicating the number of matches to return in case of more than one coincidence.}

\item{jwBound}{A real number between 0 and 1 that determines de lower bound for Jaro-Winkler distance.}
}
\value{
A data frame with the information about the matches found for the disease.
}
\description{
Main function for looking up a given pattern.
}
