% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printInfo.R
\name{printInfo}
\alias{printInfo}
\title{Print information in ICD-10 associated with a specified entry.}
\usage{
printInfo(
  string,
  inclusionTerms = TRUE,
  exclusionTerms = TRUE,
  tabular = "single"
)
}
\arguments{
\item{string}{A string containing the ICD-10  three or four-digit code from which to print info.}

\item{inclusionTerms}{A logical that tells whether to print inclusion terms or not.}

\item{exclusionTerms}{A logical that tells whether to print exclusion terms or not.}

\item{tabular}{A string that can take values 'single', 'simple', or full'.}
}
\value{
A data frame with the information about the matches found for the disease.
}
\description{
Print information in ICD-10 associated with a specified entry.
}
