% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookUpSingleDisease.R
\name{residualMatch}
\alias{residualMatch}
\title{Look up unspecified and SAI entries in catalog.}
\usage{
residualMatch(pattern, dfDisease, jwBound, tabular)
}
\arguments{
\item{pattern}{A string of the name of the disease to look up.}

\item{dfDisease}{A data frame containing all the matched diseases.}

\item{jwBound}{A real number between 0 and 1 that determines de lower bound for Jaro-Winkler distance.}

\item{tabular}{A string that can take values 'single', 'simple', or full'.}
}
\value{
A data frame with the information about the matches found for the disease.
}
\description{
Look up unspecified and SAI entries in catalog.
}
