% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizeCertificates.R
\name{tokenizeCertificates}
\alias{tokenizeCertificates}
\title{Tokenize file containing death certificates}
\usage{
tokenizeCertificates(df)
}
\arguments{
\item{df}{A data fram with the information from the death certificates}
}
\value{
A tokenized version of the causes of death.
}
\description{
Tokenize file containing death certificates
}
