% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookUpSingleDisease.R
\name{CleanString}
\alias{CleanString}
\title{Clean noise from string}
\usage{
CleanString(pattern)
}
\arguments{
\item{pattern}{The string whose noise will be removed (special characters, double spaces, etc.).}
}
\value{
A new version of the string in lowercase without noise.
}
\description{
Clean noise from string
}
