% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookUpSingleDisease.R
\name{ICDLookUp}
\alias{ICDLookUp}
\title{Look up disease in ICD10 catalog}
\usage{
ICDLookUp(pattern, nMatches = 1)
}
\arguments{
\item{pattern}{A string of the name of the disease to look up.}

\item{nMatches}{A number indicating the number of matches to return in case of more than one coincidence.}
}
\value{
A data frame with the information about the matches found for the disease.
}
\description{
Look up disease in ICD10 catalog
}
